feat(upload) added route /project/:id/upload to handle script at creation
