+++ apps/app.module.ts
@@ -4,7 +4,6 @@ import { AppService } from './app.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { ProjectModule } from './project/project.module';
-import { FileUploadModule } from "./file/file.upload.module";
 import { CategoryModule } from './category/category.module';
 import { SequenceModule } from './sequence/sequence.module';
 import { TagModule } from './tag/tag.module';
@@ -31,7 +30,6 @@ import { TagModule } from './tag/tag.module';
     TagModule,
     ProjectModule,
     SequenceModule,
-    FileUploadModule,
   ],
   controllers: [AppController],
   providers: [AppService],
