--- apps/file/file.upload.module.ts
@@ -1,15 +1,0 @@
-import { Module } from '@nestjs/common';
-import { FileUploadService } from "./file.upload.service";
-import { FileUploadController } from "./file.upload.controller";
-import {MulterModule} from "@nestjs/platform-express";
-
-@Module({
-    imports: [
-        MulterModule.register({
-            dest: '/home/node/app/files',
-        }),
-    ],
-    providers: [FileUploadService],
-    controllers: [FileUploadController],
-})
-export class FileUploadModule {}
