+++ apps/project/project.module.ts
@@ -3,9 +3,15 @@ import { ProjectController } from './project.controller';
 import { ProjectService } from './project.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Project } from './project.entity';
+import {MulterModule} from "@nestjs/platform-express";
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Project])],
+  imports: [
+      TypeOrmModule.forFeature([Project]),
+      MulterModule.register({
+          dest: '/home/node/app/files',
+      }),
+  ],
   controllers: [ProjectController],
   providers: [ProjectService],
   exports: [ProjectService],
