+++ apps/project/project.service.ts
@@ -4,6 +4,8 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { UpdateProjectDto } from './dto/update-project.dto';
 import { CreateProjectDto } from './dto/create-project.dto';
+import * as path from "path";
+import {loadFile} from "./project.utils";
 
 @Injectable()
 export class ProjectService {
@@ -12,8 +14,15 @@ export class ProjectService {
     private readonly projectRepository: Repository<Project>,
   ) {}
 
-  async create(body: CreateProjectDto): Promise<Project> {
-    const project: Project = await this.projectRepository.create(body);
+  async create(body: CreateProjectDto, file: Express.Multer.File): Promise<Project> {
+    let file_path = path.resolve(process.cwd());
+    let file_content = await loadFile(path.join(file_path, "files/", file.filename));
+
+    const project: Project = await this.projectRepository.create({
+      ...body,
+      html: file_content,
+    });
+
     return this.projectRepository.save(project);
   }
